+++ tests/torchtune/models/clip/test_clip_image_transform.py
@@ -10,7 +10,7 @@ import pytest
 
 import torch
 
-from torchtune.models.clip._component_builders import CLIPImageTransform
+from torchtune.models.clip._transforms import CLIPImageTransform
 
 
 class TestPipelines:
